refactoring - kvs19

+++ knowledge_checker/deploy/migrations/postgres/1749784566_sessions_table.up.sql
@@ -2,8 +2,8 @@ BEGIN;
 
 CREATE TABLE IF NOT EXISTS kvs.sessions (
     id SERIAL PRIMARY KEY,
-    session_id BIGSERIAL,
-    user_id SERIAL NOT NULL,
+    session_id NUMERIC,
+    user_id NUMERIC NOT NULL,
     state VARCHAR(255) NOT NULL, 
     topics TEXT[] NOT NULL,
     questions INTEGER[],
